requirements.txt
opencv-python
deepface
numpy
pandas
pillow
tk
threading
datetime
